+++ src/main/java/com/repositories/ITemplatesRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ITemplatesRepository extends JpaRepository<Templates, Long> {
+public interface ITemplatesRepository extends JpaRepository<Templates, Integer> {
     List<Templates> findByProductIn(List<Product> products);
 
     List<Templates> findByCustomerIn(List<Customer> customers);
